LIB-1 | implement library app

+++ frontend/src/app/components/user/user-list/user-list.component.ts
@@ -3,6 +3,7 @@ import { User } from '../../../models/user.model';
 import { UserService } from '../../../services/user.service';
 import { CommonModule } from '@angular/common';
 import { RouterLink } from '@angular/router';
+import { AuthService } from '../../../services/auth.service';
 
 @Component({
     selector: 'app-user-list',
@@ -14,20 +15,24 @@ import { RouterLink } from '@angular/router';
 export class UserListComponent implements OnInit {
     users: User[] = [];
 
-    constructor(private userService: UserService) {}
+    constructor(private userService: UserService, private authService: AuthService) {}
 
     ngOnInit(): void {
         this.loadUsers();
     }
 
     loadUsers(): void {
-        this.userService.getAllUsers().subscribe(data => {
-            this.users = data;
-        });
+        this.userService.getAllUsers()
+            .subscribe(data => this.users = data);
     }
 
     deleteUser(id: number | undefined): void {
         if (!id) return;
+        let currentUser = this.authService.getLoggedUser();
+        if (currentUser?.id === id) {
+            alert('Current logged in user cannot be deleted');
+            return;
+        }
         if (confirm('Are you sure you want to delete this user?')) {
             this.userService.deleteUser(id)
                 .subscribe({
